Move Set into objects/full/set.rs
