+++ crates/fj-kernel/src/objects/full/mod.rs
@@ -1,6 +1,7 @@
 pub mod cycle;
 pub mod edge;
 pub mod face;
+pub mod set;
 pub mod shell;
 pub mod sketch;
 pub mod solid;
